+++ src/main/java/org/apache/cassandra/sidecar/config/CassandraInputValidationConfiguration.java
@@ -18,181 +18,31 @@
 
 package org.apache.cassandra.sidecar.config;
 
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.HashSet;
 import java.util.Set;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import org.apache.cassandra.sidecar.common.DataObjectBuilder;
 
 /**
  * Encapsulate configuration values for validation properties used for Cassandra inputs
  */
-public class CassandraInputValidationConfiguration
+public interface CassandraInputValidationConfiguration
 {
-    public static final String FORBIDDEN_KEYSPACES_PROPERTY = "forbidden_keyspaces";
-    public static final Set<String> DEFAULT_FORBIDDEN_KEYSPACES =
-    Collections.unmodifiableSet(new HashSet<>(Arrays.asList("system_schema",
-                                                            "system_traces",
-                                                            "system_distributed",
-                                                            "system",
-                                                            "system_auth",
-                                                            "system_views",
-                                                            "system_virtual_schema")));
-    public static final String ALLOWED_CHARS_FOR_DIRECTORY_PROPERTY = "allowed_chars_for_directory";
-    public static final String DEFAULT_ALLOWED_CHARS_FOR_DIRECTORY = "[a-zA-Z0-9_-]+";
-    public static final String ALLOWED_CHARS_FOR_COMPONENT_NAME_PROPERTY = "allowed_chars_for_component_name";
-    public static final String DEFAULT_ALLOWED_CHARS_FOR_COMPONENT_NAME =
-    "[a-zA-Z0-9_-]+(.db|.cql|.json|.crc32|TOC.txt)";
-    public static final String ALLOWED_CHARS_FOR_RESTRICTED_COMPONENT_NAME_PROPERTY =
-    "allowed_chars_for_restricted_component_name";
-    public static final String DEFAULT_ALLOWED_CHARS_FOR_RESTRICTED_COMPONENT_NAME = "[a-zA-Z0-9_-]+(.db|TOC.txt)";
-
-    @JsonProperty(FORBIDDEN_KEYSPACES_PROPERTY)
-    protected final Set<String> forbiddenKeyspaces;
-
-    @JsonProperty(value = ALLOWED_CHARS_FOR_DIRECTORY_PROPERTY, defaultValue = DEFAULT_ALLOWED_CHARS_FOR_DIRECTORY)
-    protected final String allowedPatternForDirectory;
-
-    @JsonProperty(value = ALLOWED_CHARS_FOR_COMPONENT_NAME_PROPERTY,
-    defaultValue = DEFAULT_ALLOWED_CHARS_FOR_COMPONENT_NAME)
-    protected final String allowedPatternForComponentName;
-
-    @JsonProperty(value = ALLOWED_CHARS_FOR_RESTRICTED_COMPONENT_NAME_PROPERTY,
-    defaultValue = DEFAULT_ALLOWED_CHARS_FOR_RESTRICTED_COMPONENT_NAME)
-    protected final String allowedPatternForRestrictedComponentName;
-
-    public CassandraInputValidationConfiguration()
-    {
-        forbiddenKeyspaces = DEFAULT_FORBIDDEN_KEYSPACES;
-        allowedPatternForDirectory = DEFAULT_ALLOWED_CHARS_FOR_DIRECTORY;
-        allowedPatternForComponentName = DEFAULT_ALLOWED_CHARS_FOR_COMPONENT_NAME;
-        allowedPatternForRestrictedComponentName = DEFAULT_ALLOWED_CHARS_FOR_RESTRICTED_COMPONENT_NAME;
-    }
-
-    protected CassandraInputValidationConfiguration(Builder builder)
-    {
-        forbiddenKeyspaces = builder.forbiddenKeyspaces;
-        allowedPatternForDirectory = builder.allowedPatternForDirectory;
-        allowedPatternForComponentName = builder.allowedPatternForComponentName;
-        allowedPatternForRestrictedComponentName = builder.allowedPatternForRestrictedComponentName;
-    }
-
     /**
      * @return a set of forbidden keyspaces
      */
-    @JsonProperty(FORBIDDEN_KEYSPACES_PROPERTY)
-    public Set<String> forbiddenKeyspaces()
-    {
-        return forbiddenKeyspaces;
-    }
+    Set<String> forbiddenKeyspaces();
 
     /**
      * @return a regular expression for an allowed pattern for directory names
      * (i.e. keyspace directory name or table directory name)
      */
-    @JsonProperty(value = ALLOWED_CHARS_FOR_DIRECTORY_PROPERTY, defaultValue = DEFAULT_ALLOWED_CHARS_FOR_DIRECTORY)
-    public String allowedPatternForDirectory()
-    {
-        return allowedPatternForDirectory;
-    }
+    String allowedPatternForDirectory();
 
     /**
      * @return a regular expression for an allowed pattern for component names
      */
-    @JsonProperty(value = ALLOWED_CHARS_FOR_COMPONENT_NAME_PROPERTY,
-    defaultValue = DEFAULT_ALLOWED_CHARS_FOR_COMPONENT_NAME)
-    public String allowedPatternForComponentName()
-    {
-        return allowedPatternForComponentName;
-    }
+    String allowedPatternForComponentName();
 
     /**
      * @return a regular expression to an allowed pattern for a subset of component names
      */
-    @JsonProperty(value = ALLOWED_CHARS_FOR_RESTRICTED_COMPONENT_NAME_PROPERTY,
-    defaultValue = DEFAULT_ALLOWED_CHARS_FOR_RESTRICTED_COMPONENT_NAME)
-    public String allowedPatternForRestrictedComponentName()
-    {
-        return allowedPatternForRestrictedComponentName;
-    }
-
-    public static Builder builder()
-    {
-        return new Builder();
-    }
-
-    /**
-     * {@code CassandraInputValidationConfiguration} builder static inner class.
-     */
-    public static class Builder implements DataObjectBuilder<Builder, CassandraInputValidationConfiguration>
-    {
-        private Set<String> forbiddenKeyspaces = DEFAULT_FORBIDDEN_KEYSPACES;
-        private String allowedPatternForDirectory = DEFAULT_ALLOWED_CHARS_FOR_DIRECTORY;
-        private String allowedPatternForComponentName = DEFAULT_ALLOWED_CHARS_FOR_COMPONENT_NAME;
-        private String allowedPatternForRestrictedComponentName = DEFAULT_ALLOWED_CHARS_FOR_RESTRICTED_COMPONENT_NAME;
-
-        protected Builder()
-        {
-        }
-
-        /**
-         * Sets the {@code forbiddenKeyspaces} and returns a reference to this Builder enabling method chaining.
-         *
-         * @param forbiddenKeyspaces the {@code forbiddenKeyspaces} to set
-         * @return a reference to this Builder
-         */
-        public Builder forbiddenKeyspaces(Set<String> forbiddenKeyspaces)
-        {
-            return override(b -> b.forbiddenKeyspaces = forbiddenKeyspaces);
-        }
-
-        /**
-         * Sets the {@code allowedPatternForDirectory} and returns a reference to this Builder enabling method chaining.
-         *
-         * @param allowedPatternForDirectory the {@code allowedPatternForDirectory} to set
-         * @return a reference to this Builder
-         */
-        public Builder allowedPatternForDirectory(String allowedPatternForDirectory)
-        {
-            return override(b -> b.allowedPatternForDirectory = allowedPatternForDirectory);
-        }
-
-        /**
-         * Sets the {@code allowedPatternForComponentName} and returns a reference to this Builder enabling method
-         * chaining.
-         *
-         * @param allowedPatternForComponentName the {@code allowedPatternForComponentName} to set
-         * @return a reference to this Builder
-         */
-        public Builder allowedPatternForComponentName(String allowedPatternForComponentName)
-        {
-            return override(b -> b.allowedPatternForComponentName = allowedPatternForComponentName);
-        }
-
-        /**
-         * Sets the {@code allowedPatternForRestrictedComponentName} and returns a reference to this Builder
-         * enabling method chaining.
-         *
-         * @param allowedPatternForRestrictedComponentName the {@code allowedPatternForRestrictedComponentName} to set
-         * @return a reference to this Builder
-         */
-        public Builder allowedPatternForRestrictedComponentName(String allowedPatternForRestrictedComponentName)
-        {
-            return override(b -> b.allowedPatternForRestrictedComponentName = allowedPatternForRestrictedComponentName);
-        }
-
-        /**
-         * Returns a {@code CassandraInputValidationConfiguration} built from the parameters previously set.
-         *
-         * @return a {@code CassandraInputValidationConfiguration} built with parameters of this
-         * {@code CassandraInputValidationConfiguration.Builder}
-         */
-        @Override
-        public CassandraInputValidationConfiguration build()
-        {
-            return new CassandraInputValidationConfiguration(this);
-        }
-    }
+    String allowedPatternForRestrictedComponentName();
 }
