+++ src/main/java/org/apache/cassandra/sidecar/config/ServiceConfiguration.java
@@ -18,132 +18,25 @@
 
 package org.apache.cassandra.sidecar.config;
 
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.Map;
-import java.util.concurrent.TimeUnit;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import org.apache.cassandra.sidecar.common.DataObjectBuilder;
-import org.jetbrains.annotations.VisibleForTesting;
 
 /**
  * Configuration for the Sidecar Service and configuration of the REST endpoints in the service
  */
-public class ServiceConfiguration
+public interface ServiceConfiguration
 {
-    public static final String HOST_PROPERTY = "host";
-    public static final String DEFAULT_HOST = "0.0.0.0";
-    public static final String PORT_PROPERTY = "port";
-    public static final int DEFAULT_PORT = 9043;
-    public static final String REQUEST_IDLE_TIMEOUT_MILLIS_PROPERTY = "request_idle_timeout_millis";
-    public static final int DEFAULT_REQUEST_IDLE_TIMEOUT_MILLIS = 300000;
-    public static final String REQUEST_TIMEOUT_MILLIS_PROPERTY = "request_timeout_millis";
-    public static final long DEFAULT_REQUEST_TIMEOUT_MILLIS = 300000L;
-    public static final String ALLOWABLE_SKEW_IN_MINUTES_PROPERTY = "allowable_time_skew_in_minutes";
-    public static final int DEFAULT_ALLOWABLE_SKEW_IN_MINUTES = 60;
-    public static final String THROTTLE_PROPERTY = "throttle";
-    public static final String SSTABLE_UPLOAD_PROPERTY = "sstable_upload";
-    public static final String SSTABLE_IMPORT_PROPERTY = "sstable_import";
-    public static final String WORKER_POOLS_PROPERTY = "worker_pools";
-
-    public static final String SERVICE_POOL = "service";
-    public static final String INTERNAL_POOL = "internal";
-    protected static final Map<String, WorkerPoolConfiguration> DEFAULT_WORKER_POOLS_CONFIGURATION
-    = Collections.unmodifiableMap(new HashMap<String, WorkerPoolConfiguration>()
-    {{
-        put(SERVICE_POOL, WorkerPoolConfiguration.builder()
-                                                 .workerPoolName("sidecar-worker-pool")
-                                                 .workerPoolSize(20)
-                                                 .workerMaxExecutionTimeMillis(TimeUnit.SECONDS.toMillis(60))
-                                                 .build());
-
-
-        put(INTERNAL_POOL, WorkerPoolConfiguration.builder()
-                                                  .workerPoolName("sidecar-internal-worker-pool")
-                                                  .workerPoolSize(20)
-                                                  .workerMaxExecutionTimeMillis(TimeUnit.MINUTES.toMillis(15))
-                                                  .build());
-    }});
-
-
-    @JsonProperty(value = HOST_PROPERTY, defaultValue = DEFAULT_HOST)
-    protected final String host;
-
-    @JsonProperty(value = PORT_PROPERTY, defaultValue = DEFAULT_PORT + "")
-    private final int port;
-
-    @JsonProperty(value = REQUEST_IDLE_TIMEOUT_MILLIS_PROPERTY, defaultValue = DEFAULT_REQUEST_IDLE_TIMEOUT_MILLIS + "")
-    private final int requestIdleTimeoutMillis;
-
-    @JsonProperty(value = REQUEST_TIMEOUT_MILLIS_PROPERTY, defaultValue = DEFAULT_REQUEST_TIMEOUT_MILLIS + "")
-    private final long requestTimeoutMillis;
-
-    @JsonProperty(value = ALLOWABLE_SKEW_IN_MINUTES_PROPERTY, defaultValue = DEFAULT_ALLOWABLE_SKEW_IN_MINUTES + "")
-    private final int allowableSkewInMinutes;
-
-    @JsonProperty(value = THROTTLE_PROPERTY, required = true)
-    private final ThrottleConfiguration throttleConfiguration;
-
-    @JsonProperty(value = SSTABLE_UPLOAD_PROPERTY, required = true)
-    private final SSTableUploadConfiguration ssTableUploadConfiguration;
-
-    @JsonProperty(value = SSTABLE_IMPORT_PROPERTY, required = true)
-    private final SSTableImportConfiguration ssTableImportConfiguration;
-
-    @JsonProperty(value = WORKER_POOLS_PROPERTY, required = true)
-    private final Map<String, ? extends WorkerPoolConfiguration> workerPoolsConfiguration;
-
-    public ServiceConfiguration()
-    {
-        host = DEFAULT_HOST;
-        port = DEFAULT_PORT;
-        requestIdleTimeoutMillis = DEFAULT_REQUEST_IDLE_TIMEOUT_MILLIS;
-        requestTimeoutMillis = DEFAULT_REQUEST_TIMEOUT_MILLIS;
-        allowableSkewInMinutes = DEFAULT_ALLOWABLE_SKEW_IN_MINUTES;
-        throttleConfiguration = new ThrottleConfiguration();
-        ssTableUploadConfiguration = new SSTableUploadConfiguration();
-        ssTableImportConfiguration = new SSTableImportConfiguration();
-        workerPoolsConfiguration = DEFAULT_WORKER_POOLS_CONFIGURATION;
-    }
-
-    protected ServiceConfiguration(Builder<?> builder)
-    {
-        host = builder.host;
-        port = builder.port;
-        requestIdleTimeoutMillis = builder.requestIdleTimeoutMillis;
-        requestTimeoutMillis = builder.requestTimeoutMillis;
-        allowableSkewInMinutes = builder.allowableSkewInMinutes;
-        throttleConfiguration = builder.throttleConfiguration;
-        ssTableUploadConfiguration = builder.ssTableUploadConfiguration;
-        ssTableImportConfiguration = builder.ssTableImportConfiguration;
-        if (builder.workerPoolsConfiguration == null || builder.workerPoolsConfiguration.isEmpty())
-        {
-            workerPoolsConfiguration = DEFAULT_WORKER_POOLS_CONFIGURATION;
-        }
-        else
-        {
-            workerPoolsConfiguration = builder.workerPoolsConfiguration;
-        }
-    }
+    String SERVICE_POOL = "service";
+    String INTERNAL_POOL = "internal";
 
     /**
      * Sidecar's HTTP REST API listen address
      */
-    @JsonProperty(value = HOST_PROPERTY, defaultValue = DEFAULT_HOST)
-    public String host()
-    {
-        return host;
-    }
+    String host();
 
     /**
      * @return Sidecar's HTTP REST API port
      */
-    @JsonProperty(value = PORT_PROPERTY, defaultValue = DEFAULT_PORT + "")
-    public int port()
-    {
-        return port;
-    }
+    int port();
 
     /**
      * Determines if a connection will timeout and be closed if no data is received nor sent within the timeout.
@@ -151,261 +44,51 @@ public class ServiceConfiguration
      *
      * @return the configured idle timeout value
      */
-    @JsonProperty(value = REQUEST_IDLE_TIMEOUT_MILLIS_PROPERTY, defaultValue = DEFAULT_REQUEST_IDLE_TIMEOUT_MILLIS + "")
-    public int requestIdleTimeoutMillis()
-    {
-        return requestIdleTimeoutMillis;
-    }
+    int requestIdleTimeoutMillis();
 
     /**
      * Determines if a response will timeout if the response has not been written after a certain time.
      */
-    @JsonProperty(value = REQUEST_TIMEOUT_MILLIS_PROPERTY, defaultValue = DEFAULT_REQUEST_TIMEOUT_MILLIS + "")
-    public long requestTimeoutMillis()
-    {
-        return requestTimeoutMillis;
-    }
+    long requestTimeoutMillis();
 
     /**
      * @return the maximum time skew allowed between the server and the client
      */
-    @JsonProperty(value = ALLOWABLE_SKEW_IN_MINUTES_PROPERTY, defaultValue = DEFAULT_ALLOWABLE_SKEW_IN_MINUTES + "")
-    public int allowableSkewInMinutes()
-    {
-        return allowableSkewInMinutes;
-    }
+    int allowableSkewInMinutes();
 
     /**
      * @return the throttling configuration
      */
-    @JsonProperty(value = THROTTLE_PROPERTY, required = true)
-    public ThrottleConfiguration throttleConfiguration()
-    {
-        return throttleConfiguration;
-    }
+    ThrottleConfiguration throttleConfiguration();
 
     /**
      * @return the configuration for SSTable component uploads on this service
      */
-    @JsonProperty(value = SSTABLE_UPLOAD_PROPERTY, required = true)
-    public SSTableUploadConfiguration ssTableUploadConfiguration()
-    {
-        return ssTableUploadConfiguration;
-    }
+    SSTableUploadConfiguration ssTableUploadConfiguration();
 
     /**
      * @return the configuration for the SSTable Import functionality
      */
-    @JsonProperty(value = SSTABLE_IMPORT_PROPERTY, required = true)
-    public SSTableImportConfiguration ssTableImportConfiguration()
-    {
-        return ssTableImportConfiguration;
-    }
+    SSTableImportConfiguration ssTableImportConfiguration();
 
     /**
      * @return the configured worker pools for the service
      */
-    @JsonProperty(value = WORKER_POOLS_PROPERTY, required = true)
-    public Map<String, ? extends WorkerPoolConfiguration> workerPoolsConfiguration()
-    {
-        return workerPoolsConfiguration;
-    }
+    Map<String, ? extends WorkerPoolConfiguration> workerPoolsConfiguration();
 
     /**
      * @return the configuration for the {@link #SERVICE_POOL}
      */
-    public WorkerPoolConfiguration serverWorkerPoolConfiguration()
+    default WorkerPoolConfiguration serverWorkerPoolConfiguration()
     {
-        return workerPoolsConfiguration.get(SERVICE_POOL);
+        return workerPoolsConfiguration().get(SERVICE_POOL);
     }
 
     /**
      * @return the configuration for the {@link #INTERNAL_POOL}
      */
-    public WorkerPoolConfiguration serverInternalWorkerPoolConfiguration()
+    default WorkerPoolConfiguration serverInternalWorkerPoolConfiguration()
     {
-        return workerPoolsConfiguration.get(INTERNAL_POOL);
-    }
-
-    @VisibleForTesting
-    public Builder<?> unbuild()
-    {
-        return new Builder<>(this);
-    }
-
-    public static Builder<?> builder()
-    {
-        return new Builder<>();
-    }
-
-    /**
-     * {@code ServiceConfiguration} builder static inner class.
-     */
-    public static class Builder<T extends Builder<?>> implements DataObjectBuilder<T, ServiceConfiguration>
-    {
-        protected String host = DEFAULT_HOST;
-        protected int port = DEFAULT_PORT;
-        protected int requestIdleTimeoutMillis = DEFAULT_REQUEST_IDLE_TIMEOUT_MILLIS;
-        protected long requestTimeoutMillis = DEFAULT_REQUEST_TIMEOUT_MILLIS;
-        protected int allowableSkewInMinutes = DEFAULT_ALLOWABLE_SKEW_IN_MINUTES;
-        protected ThrottleConfiguration throttleConfiguration = new ThrottleConfiguration();
-        protected SSTableUploadConfiguration ssTableUploadConfiguration = new SSTableUploadConfiguration();
-        protected SSTableImportConfiguration ssTableImportConfiguration = new SSTableImportConfiguration();
-        protected Map<String, WorkerPoolConfiguration> workerPoolsConfiguration = new HashMap<>();
-
-        protected Builder()
-        {
-        }
-
-        @SuppressWarnings("unchecked")
-        protected Builder(ServiceConfiguration serviceConfiguration)
-        {
-            host = serviceConfiguration.host;
-            port = serviceConfiguration.port;
-            requestIdleTimeoutMillis = serviceConfiguration.requestIdleTimeoutMillis;
-            requestTimeoutMillis = serviceConfiguration.requestTimeoutMillis;
-            allowableSkewInMinutes = serviceConfiguration.allowableSkewInMinutes;
-            throttleConfiguration = serviceConfiguration.throttleConfiguration;
-            ssTableUploadConfiguration = serviceConfiguration.ssTableUploadConfiguration;
-            ssTableImportConfiguration = serviceConfiguration.ssTableImportConfiguration;
-            workerPoolsConfiguration = (Map<String, WorkerPoolConfiguration>)
-                                       serviceConfiguration.workerPoolsConfiguration;
-        }
-
-        /**
-         * Sets the {@code host} and returns a reference to this Builder enabling method chaining.
-         *
-         * @param host the {@code host} to set
-         * @return a reference to this Builder
-         */
-        public T host(String host)
-        {
-            return override(b -> b.host = host);
-        }
-
-        /**
-         * Sets the {@code port} and returns a reference to this Builder enabling method chaining.
-         *
-         * @param port the {@code port} to set
-         * @return a reference to this Builder
-         */
-        public T port(int port)
-        {
-            return override(b -> b.port = port);
-        }
-
-        /**
-         * Sets the {@code requestIdleTimeoutMillis} and returns a reference to this Builder enabling method chaining.
-         *
-         * @param requestIdleTimeoutMillis the {@code requestIdleTimeoutMillis} to set
-         * @return a reference to this Builder
-         */
-        public T requestIdleTimeoutMillis(int requestIdleTimeoutMillis)
-        {
-            return override(b -> b.requestIdleTimeoutMillis = requestIdleTimeoutMillis);
-        }
-
-        /**
-         * Sets the {@code requestTimeoutMillis} and returns a reference to this Builder enabling method chaining.
-         *
-         * @param requestTimeoutMillis the {@code requestTimeoutMillis} to set
-         * @return a reference to this Builder
-         */
-        public T requestTimeoutMillis(long requestTimeoutMillis)
-        {
-            return override(b -> b.requestTimeoutMillis = requestTimeoutMillis);
-        }
-
-        /**
-         * Sets the {@code allowableSkewInMinutes} and returns a reference to this Builder enabling method chaining.
-         *
-         * @param allowableSkewInMinutes the {@code allowableSkewInMinutes} to set
-         * @return a reference to this Builder
-         */
-        public T allowableSkewInMinutes(int allowableSkewInMinutes)
-        {
-            return override(b -> b.allowableSkewInMinutes = allowableSkewInMinutes);
-        }
-
-        /**
-         * Sets the {@code throttleConfiguration} and returns a reference to this Builder enabling method chaining.
-         *
-         * @param throttleConfiguration the {@code throttleConfiguration} to set
-         * @return a reference to this Builder
-         */
-        public T throttleConfiguration(ThrottleConfiguration throttleConfiguration)
-        {
-            return override(b -> b.throttleConfiguration = throttleConfiguration);
-        }
-
-        /**
-         * Sets the {@code ssTableUploadConfiguration} and returns a reference to this Builder enabling method chaining.
-         *
-         * @param ssTableUploadConfiguration the {@code ssTableUploadConfiguration} to set
-         * @return a reference to this Builder
-         */
-        public T ssTableUploadConfiguration(SSTableUploadConfiguration ssTableUploadConfiguration)
-        {
-            return override(b -> b.ssTableUploadConfiguration = ssTableUploadConfiguration);
-        }
-
-        /**
-         * Sets the {@code ssTableImportConfiguration} and returns a reference to this Builder enabling method chaining.
-         *
-         * @param ssTableImportConfiguration the {@code ssTableImportConfiguration} to set
-         * @return a reference to this Builder
-         */
-        public T ssTableImportConfiguration(SSTableImportConfiguration ssTableImportConfiguration)
-        {
-            return override(b -> b.ssTableImportConfiguration = ssTableImportConfiguration);
-        }
-
-        /**
-         * Sets the {@code workerPoolsConfiguration} and returns a reference to this Builder enabling method chaining.
-         *
-         * @param workerPoolsConfiguration the {@code workerPoolsConfiguration} to set
-         * @return a reference to this Builder
-         */
-        public T workerPoolsConfiguration(Map<String, WorkerPoolConfiguration> workerPoolsConfiguration)
-        {
-            return override(b -> b.workerPoolsConfiguration = workerPoolsConfiguration);
-        }
-
-        /**
-         * Sets the {@link #SERVICE_POOL} configuration to {@code workerPoolConfiguration} and returns a
-         * reference to this Builder enabling method chaining.
-         *
-         * @param workerPoolConfiguration the {@code workerPoolConfiguration} to set for the {@link #SERVICE_POOL}
-         * @return a reference to this Builder
-         */
-        public T servicePoolConfiguration(WorkerPoolConfiguration workerPoolConfiguration)
-        {
-            workerPoolsConfiguration.put(SERVICE_POOL, workerPoolConfiguration);
-            return self();
-        }
-
-        /**
-         * Sets the {@link #INTERNAL_POOL} configuration to {@code workerPoolConfiguration} and returns a
-         * reference to this Builder enabling method chaining.
-         *
-         * @param workerPoolConfiguration the {@code workerPoolConfiguration} to set for the {@link #INTERNAL_POOL}
-         * @return a reference to this Builder
-         */
-        public T internalPoolConfiguration(WorkerPoolConfiguration workerPoolConfiguration)
-        {
-            workerPoolsConfiguration.put(INTERNAL_POOL, workerPoolConfiguration);
-            return self();
-        }
-
-        /**
-         * Returns a {@code ServiceConfiguration} built from the parameters previously set.
-         *
-         * @return a {@code ServiceConfiguration} built with parameters of this {@code ServiceConfiguration.Builder}
-         */
-        @Override
-        public ServiceConfiguration build()
-        {
-            return new ServiceConfiguration(this);
-        }
+        return workerPoolsConfiguration().get(INTERNAL_POOL);
     }
 }
