+++ src/test/java/org/apache/cassandra/sidecar/snapshots/SnapshotSearchTest.java
@@ -38,7 +38,7 @@ import io.vertx.junit5.VertxExtension;
 import io.vertx.junit5.VertxTestContext;
 import org.apache.cassandra.sidecar.cluster.InstancesConfig;
 import org.apache.cassandra.sidecar.concurrent.ExecutorPools;
-import org.apache.cassandra.sidecar.config.ServiceConfiguration;
+import org.apache.cassandra.sidecar.config.impl.ServiceConfigurationImpl;
 import org.apache.cassandra.sidecar.utils.CassandraInputValidator;
 
 import static org.apache.cassandra.sidecar.snapshots.SnapshotUtils.mockInstancesConfig;
@@ -67,7 +67,7 @@ public class SnapshotSearchTest
         InstancesConfig mockInstancesConfig = mockInstancesConfig(rootDir);
 
         CassandraInputValidator validator = new CassandraInputValidator();
-        ExecutorPools executorPools = new ExecutorPools(vertx, ServiceConfiguration.builder().build());
+        ExecutorPools executorPools = new ExecutorPools(vertx, ServiceConfigurationImpl.builder().build());
         instance = new SnapshotPathBuilder(vertx, mockInstancesConfig, validator, executorPools);
     }
